rough outline
